requirements.txt
nextcord==2.0.0a6
python-dotenv==1.0.0
